requirements.txt
fastapi
uvicorn[standard]
authlib
httpx
python-dotenv
python-jose
passlib[bcrypt]
python-multipart
itsdangerous
xmltodict

